requirements.txt
asgiref==3.4.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
certifi==2021.10.8
cffi==1.15.0
cryptography==36.0.0
distlib==0.3.3
Django==3.2.8
djangorestframework==3.12.4
drf-jwt==1.19.1
ecdsa==0.17.0
filelock==3.3.1
mysqlclient==2.0.3
pipenv==2021.5.29
platformdirs==2.4.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
PyJWT==1.7.1
pytz==2021.3
rsa==4.8
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
virtualenv==20.9.0
virtualenv-clone==0.5.7
